Migrating from mongoose to sequelize, introducing types

+++ server/lib/server.js
@@ -27,9 +27,8 @@ database.connect(function(err) {
 			},
 
 			setPadId : function(padId) {
-				if(socket.padId != null) {
+				if(typeof padId != "string" || socket.padId != null)
 					return;
-				}
 
 				socket.padId = true;
 
@@ -43,6 +42,7 @@ database.connect(function(err) {
 
 					_sendData(socket, "padData", null, data);
 					_sendStreamData(socket, "view", database.getViews(socket.padId));
+					_sendStreamData(socket, "type", database.getTypes(socket.padId));
 					_sendStreamData(socket, "line", database.getPadLines(socket.padId));
 
 					if(socket.bbox) { // In case bbox is set while fetching pad data
@@ -53,6 +53,9 @@ database.connect(function(err) {
 			},
 
 			updateBbox : function(bbox) {
+				if(!utils.stripObject(bbox, { top: "number", left: "number", bottom: "number", right: "number", zoom: "number" }))
+					return;
+
 				var bboxWithExcept = utils.extend({ }, bbox);
 				if(socket.bbox && bbox.zoom == socket.bbox.zoom)
 					bboxWithExcept.except = socket.bbox;
@@ -71,78 +74,139 @@ database.connect(function(err) {
 			},
 
 			editPad : function(data, callback) {
+				if(!utils.stripObject(data, { name: "string", defaultViewId: "number" }))
+					return callback("Invalid parameters.");
+
 				if(!socket.writable)
-					return callback(new Error("In read-only mode."));
+					return callback("In read-only mode.");
 				
 				database.updatePadData(socket.padId, data, callback);
 			},
 
 			addMarker : function(data, callback) {
+				if(!utils.stripObject(data, { lat: "number", lon: "number", name: "string", colour: "string", typeId: "number" } ))
+					return callback("Invalid parameters.");
+
 				if(!socket.writable)
-					return callback(new Error("In read-only mode."));
+					return callback("In read-only mode.");
 				
 				database.createMarker(socket.padId, data, callback);
 			},
 
 			editMarker : function(data, callback) {
+				if(!utils.stripObject(data, { id: "number", lat: "number", lon: "number", name: "string", colour: "string" }))
+					return callback("Invalid parameters.");
+
 				if(!socket.writable)
-					return callback(new Error("In read-only mode."));
+					return callback("In read-only mode.");
 				
 				database.updateMarker(data.id, data, callback);
 			},
 
 			deleteMarker : function(data, callback) {
+				if(!utils.stripObject(data, { id: "number" }))
+					return callback("Invalid parameters.");
+
 				if(!socket.writable)
-					return callback(new Error("In read-only mode."));
+					return callback("In read-only mode.");
 				
 				database.deleteMarker(data.id, callback);
 			},
 
 			addLine : function(data, callback) {
+				if(!utils.stripObject(data, { points: [ { lat: "number", lon: "number" } ], mode: "string", colour: "string", width: "number", name: "string", typeId: "number" }))
+					return callback("Invalid parameters.");
+
 				if(!socket.writable)
-					return callback(new Error("In read-only mode."));
+					return callback("In read-only mode.");
 				
 				database.createLine(socket.padId, data, callback);
 			},
 
 			editLine : function(data, callback) {
+				if(!utils.stripObject(data, { id: "number", points: [ { lat: "number", lon: "number" } ], mode: "string", colour: "string", width: "number", name: "string" }))
+					return callback("Invalid parameters.");
+
 				if(!socket.writable)
-					return callback(new Error("In read-only mode."));
+					return callback("In read-only mode.");
 				
 				database.updateLine(data.id, data, callback);
 			},
 
 			deleteLine : function(data, callback) {
+				if(!utils.stripObject(data, { id: "number" }))
+					return callback("Invalid parameters.");
+
 				if(!socket.writable)
-					return callback(new Error("In read-only mode."));
+					return callback("In read-only mode.");
 				
 				database.deleteLine(data.id, callback);
 			},
 
 			addView : function(data, callback) {
+				if(!utils.stripObject(data, { name: "string", baseLayer: "string", layers: [ "string" ], top: "number", left: "number", right: "number", bottom: "number" }))
+					return callback("Invalid parameters.");
+
 				if(!socket.writable)
-					return callback(new Error("In read-only mode."));
+					return callback("In read-only mode.");
 
 				database.createView(socket.padId, data, callback);
 			},
 
 			editView : function(data, callback) {
+				if(!utils.stripObject(data, { id: "number", baseLayer: "string", layers: [ "string" ], top: "number", left: "number", right: "number", bottom: "number" }))
+					return callback("Invalid parameters.");
+
 				if(!socket.writable)
-					return callback(new Error("In read-only mode."));
+					return callback("In read-only mode.");
 
 				database.updateView(data.id, data, callback);
 			},
 
 			deleteView : function(data, callback) {
+				if(!utils.stripObject(data, { id: "number" }))
+					return callback("Invalid parameters.");
+
 				if(!socket.writable)
-					return callback(new Error("In read-only mode."));
+					return callback("In read-only mode.");
 
 				database.deleteView(data.id, callback);
 			},
 
+			addType : function(data, callback) {
+				// TODO: Strip object
+
+				if(!socket.writable)
+					return callback("In read-only mode.");
+
+				database.createType(socket.padId, data, callback);
+			},
+
+			editType : function(data, callback) {
+				// TODO: Strip object
+
+				if(!socket.writable)
+					return callback("In read-only mode.");
+
+				database.updateType(data.id, data, callback);
+			},
+
+			deleteType : function(data, callback) {
+				if(!utils.stripObject(data, { id: "number" }))
+					return callback("Invalid parameters.");
+
+				if(!socket.writable)
+					return callback("In read-only mode.");
+
+				database.deleteType(data.id, callback);
+			}/*,
+
 			copyPad : function(data, callback) {
+				if(!utils.stripObject(data, { toId: "string" }))
+					return callback("Invalid parameters.");
+
 				database.copyPad(socket.padId, data.toId, callback);
-			}
+			}*/
 		};
 
 		for(var i in handlers)
